pyproject.toml
[project]
name = "cve-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0.1",
]
